requirements.txt
streamlit
tensorflow
joblib
numpy
Pillow
